AMDGPU: Don't assume CFG order in SIInsertSkips
This pass has an optimization to end the program when all the threads
are killed. It tried to only do this in uniform control flow by tracking
non-uniform if statements, but the code for tracking ifs assumed that
the ordering of blocks followed the if-nesting even though this pass ran
after the block-reordering pass which can mess it up. Visiting the
blocks in reverse post-order should obey the nesting. We no longer visit
the created blocks, but that shouldn't matter.

The pass still doesn't take non-uniform loops into account, but at least
this is a start to fixing it.